+++ packages/website-eslint/src/mock/semver.js
@@ -1,5 +1,5 @@
-import satisfies from 'semver/functions/satisfies';
 import major from 'semver/functions/major';
+import satisfies from 'semver/functions/satisfies';
 
 // just in case someone adds a import * as semver usage
 export { satisfies, major };